Remove dependency on creating a tooltip object
Also, exit the Filter function faster to improve performance if it's disabled, or if the `equipSlot` field is invalid.

I noticed that when AllTheThings refreshes its collection, creating Tooltips from items takes a LONG time, so removing the dependency on the tooltip allows this filter to not chug after ATT refreshes its collection.